requirements.txt
tensorflow
numpy
matplotlib
opencv-python
tensorflow-object-detection-api
time